+++ apix-impl/src/main/java/eu/xenit/apix/alfresco/search/SearchFacetsServiceImpl.java
@@ -38,9 +38,6 @@ import org.alfresco.service.namespace.QName;
 import org.alfresco.util.Pair;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.BeansException;
-import org.springframework.beans.factory.BeanFactory;
-import org.springframework.beans.factory.BeanFactoryAware;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -48,15 +45,13 @@ import org.springframework.stereotype.Component;
 // Note: To *temporarily* get working syntax highlighting and IDE features in this file: change the root
 // build.gradle to have alfresco_4_version = "5.0.d" rather than "4.2.f". Change it back before committing.
 @Component
-public class SearchFacetsServiceImpl implements SearchFacetsService, BeanFactoryAware, InitializingBean {
+public class SearchFacetsServiceImpl implements SearchFacetsService {
 
     private static final Logger logger = LoggerFactory.getLogger(SearchFacetsServiceImpl.class);
 
-    // Beanfactory added to replace alfresco's ServiceRegistry.getFacetLabelDisplayHandlerRegistry
-    // (Due to alfresco planning to remove the class from the public api)
-    private BeanFactory beanFactory;
-
+    @Autowired
     private FacetLabelDisplayHandlerRegistry facetLabelDisplayHandlerRegistry;
+    @Autowired
     private SolrFacetHelper solrFacetHelper;
     private DictionaryService dictionaryService;
     private NodeService nodeService;
@@ -73,18 +68,13 @@ public class SearchFacetsServiceImpl implements SearchFacetsService, BeanFactory
         nodeService = serviceRegistry.getNodeService();
     }
 
-    @Override
-    public void setBeanFactory(BeanFactory beanFactory) throws BeansException {
-        this.beanFactory = beanFactory;
-    }
-
-    @Override
-    public void afterPropertiesSet() throws Exception {
-        final String solrFacetHelperBeanName = "facet.solrFacetHelper";
-        solrFacetHelper = (SolrFacetHelper) beanFactory.getBean(solrFacetHelperBeanName);
-        final String facetLabelDisplayHandlerRegistryBeanName = "facet.facetLabelDisplayHandlerRegistry";
-        facetLabelDisplayHandlerRegistry = (FacetLabelDisplayHandlerRegistry) beanFactory.getBean(facetLabelDisplayHandlerRegistryBeanName);
-    }
+//    @Override
+//    public void afterPropertiesSet() throws Exception {
+//        final String solrFacetHelperBeanName = "facet.solrFacetHelper";
+//        solrFacetHelper = (SolrFacetHelper) beanFactory.getBean(solrFacetHelperBeanName);
+//        final String facetLabelDisplayHandlerRegistryBeanName = "facet.facetLabelDisplayHandlerRegistry";
+//        facetLabelDisplayHandlerRegistry = (FacetLabelDisplayHandlerRegistry) beanFactory.getBean(facetLabelDisplayHandlerRegistryBeanName);
+//    }
 
     public FacetLabelDisplayHandlerRegistry getFacetLabelDisplayHandlerRegistry() {
         return facetLabelDisplayHandlerRegistry;