+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 ### Added
 
 ### Changed
-* [ALFREDAPI-378](https://xenitsupport.jira.com/browse/ALFREDAPI-378): Replace depricated calls to serviceregistry
+* [ALFREDAPI-378](https://xenitsupport.jira.com/browse/ALFREDAPI-378): Replace deprecated calls to serviceregistry
 
 ### Deleted